+++ example/box.dart
@@ -36,4 +36,6 @@ void main() {
   // tampered with or there was otherwise an error.
   final decrypted = aliceBox.decrypt(encrypted);
   print(String.fromCharCodes(decrypted.plaintext));
+
+  
 }