+++ CHANGELOG.md
@@ -7,3 +7,8 @@
 - Added the `byte-length` official SHA512 test vectors.
 - Added `hashing` example
 - Allowed `personalisation` paramater to be less then 16 bytes long by zero-padding to 16 bytes.
+
+## 0.1.1-dev.1
+
+- Added the `Curve25519`'s official `DH (Diffie-Hellman)` test vector.
+- Added the `Wycheproof`'s X25519 test vectors.